requirements.txt
flask
gensim
numpy
nltk
pyspellchecker
textblob
scipy
beautifulsoup4
sockets
whois
google
requests